WIP - Working on locking

+++ src/MassTransit.RedisSagas.RedLock/RedisSagaConsumeContext.cs
@@ -3,26 +3,30 @@ using System.Threading.Tasks;
 using MassTransit.Context;
 using MassTransit.Logging;
 using MassTransit.RedisSaga;
+using MassTransit.RedisSagas.RedLock;
 using MassTransit.Util;
+using RedLock;
 using StackExchange.Redis;
 
 namespace MassTransit.RedisSagas
 {
-    public class RedisSagaConsumeContext<TSaga, TMessage> :
+    public class RedLockSagaConsumeContext<TSaga, TMessage> :
         ConsumeContextProxyScope<TMessage>,
         SagaConsumeContext<TSaga, TMessage>
         where TMessage : class
         where TSaga : class, IVersionedSaga
     {
-        static readonly ILog Log = Logger.Get<RedisSagaRepository<TSaga>>();
+        static readonly ILog Log = Logger.Get<RedLockSagaRepository<TSaga>>();
         readonly IDatabase _redisDb;
+        private readonly IRedisLockFactory _lockFactory;
         private readonly string _redisPrefix;
 
-        public RedisSagaConsumeContext(IDatabase redisDb, ConsumeContext<TMessage> context, TSaga instance, string redisPrefix = "")
+        public RedLockSagaConsumeContext(IDatabase redisDb, IRedisLockFactory lockFactory, ConsumeContext<TMessage> context, TSaga instance, string redisPrefix = "")
             : base(context)
         {
             Saga = instance;
             _redisDb = redisDb;
+            _lockFactory = lockFactory;
             _redisPrefix = redisPrefix;
         }
 
@@ -39,12 +43,24 @@ namespace MassTransit.RedisSagas
         async Task SagaConsumeContext<TSaga>.SetCompleted()
         {
             ITypedDatabase<TSaga> db = _redisDb.As<TSaga>();
-            await db.Delete(Saga.CorrelationId, _redisPrefix).ConfigureAwait(false);
+
+            using(var distLock  = _lockFactory.Create($"redislock:{Saga.CorrelationId}", TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(0.5)))
+            {
+                if (Log.IsDebugEnabled)
+                    Log.Debug($"SAGA:{TypeMetadataCache<TSaga>.ShortName}:{TypeMetadataCache<TMessage>.ShortName} Entering Lock {Saga.CorrelationId}");
+
+                if (distLock.IsAcquired)
+                {
+                    await db.Delete(Saga.CorrelationId, _redisPrefix).ConfigureAwait(false);
+                }
+
+                if (Log.IsDebugEnabled)
+                    Log.Debug($"SAGA:{TypeMetadataCache<TSaga>.ShortName}:{TypeMetadataCache<TMessage>.ShortName} Leaving Lock {Saga.CorrelationId}");
+            }
 
             IsCompleted = true;
             if (Log.IsDebugEnabled)
-                Log.DebugFormat("SAGA:{0}:{1} Removed {2}", TypeMetadataCache<TSaga>.ShortName, TypeMetadataCache<TMessage>.ShortName,
-                    Saga.CorrelationId);
+                Log.Debug($"SAGA:{TypeMetadataCache<TSaga>.ShortName}:{TypeMetadataCache<TMessage>.ShortName} Removed {Saga.CorrelationId}");
         }
 
         public TSaga Saga { get; }
